refactor: fix all remaining compilation errors for arbitrary player-ID system
- arena.rs: cast winner position to u16 instead of u8
- combat.rs: update all internal tuple type annotations from u8 to u16 for player_id fields (positions, interceptor_actions, chaff_positions, shields snapshots, AttackEvent::Melee attacker_team)
- economy.rs: add ai_player_id: u16 param to start_ai_battle; replace hardcoded players[1] indexing with progress.player(ai_player_id) / progress.player_mut(ai_player_id); pass ai_player_id to spawn_ai_army
- build_phase.rs: replace removed arena::deploy_x_range() calls with progress.player(local_player_id).deploy_zone; add ai_pid lookup and pass to start_ai_battle at both call sites
- game_over.rs: collect player_ids before creating new MatchProgress; pass next_id to BuildState::new instead of is_host bool

+++ src/arena.rs
@@ -28,7 +28,7 @@ pub fn check_match_state(units: &[Unit]) -> MatchState {
     match alive_count {
         0 => MatchState::Draw,
         1 => {
-            let winner = team_alive.iter().position(|&a| a).unwrap() as u8;
+            let winner = team_alive.iter().position(|&a| a).unwrap() as u16;
             MatchState::Winner(winner)
         }
         _ => MatchState::InProgress,
